Optimizations for older NVidia Pascal arch.  Added file output option for testing and normalization fixes for various audio level issues.  Added test audio script for varifying audio routing.  Improved Ctrl-C interrupting of client (and actions).

+++ whisper_live/server.py
@@ -682,7 +682,7 @@ class TranscriptionServer:
             transcriber = self.rest_models.get(model_name)
             if transcriber is None:
                 device = "cuda" if torch.cuda.is_available() else "cpu"
-                compute_type = "float16" if device == "cuda" else "int8"
+                compute_type = "float32" if device == "cuda" else "int8"  # modified to float32 for Nvidia Pascal arch 
                 logging.info(f"Loading REST model '{model_name}' on {device}")
                 transcriber = WhisperModel(model_name, device=device, compute_type=compute_type)
                 self.rest_models[model_name] = transcriber
